+++ main/src/main/java/net/oneandone/stool/extensions/Extensions.java
@@ -106,4 +106,12 @@ public class Extensions {
         }
         return result;
     }
+
+    public void files(FileNode dest) throws IOException {
+        for (Switch s : extensions.values()) {
+            if (s.enabled) {
+                s.extension.files(dest);
+            }
+        }
+    }
 }