+++ main/src/main/java/net/oneandone/stool/extensions/Fitnesse.java
@@ -50,4 +50,8 @@ public class Fitnesse implements Extension {
     @Override
     public void containerOpts(Stage stage, Map<String, Object> result) {
     }
+
+    @Override
+    public void files(FileNode dest) throws IOException {
+    }
 }