+++ main/src/main/java/net/oneandone/stool/extensions/Logstash.java
@@ -92,4 +92,8 @@ public class Logstash implements Extension {
     @Override
     public void containerOpts(Stage stage, Map<String, Object> result) {
     }
+
+    @Override
+    public void files(FileNode dest) throws IOException {
+    }
 }